shop was complete

+++ catalog/views.py
@@ -7,12 +7,19 @@ import openpyxl
 from io import BytesIO
 from django.core.mail import EmailMessage
 from django.conf import settings
-from rest_framework import viewsets, permissions
+from rest_framework import viewsets, permissions, generics
 from .serializers import (
     ProductSerializer, CategorySerializer, ManufacturerSerializer,
-    CartSerializer, CartItemSerializer
+    CartSerializer, CartItemSerializer, ProfileSerializer, OrderSerializer, OrderItemSerializer,Order
 )
+from django.contrib.auth import login, logout, authenticate
+from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 from django.core.paginator import Paginator
+from django.contrib.auth.forms import PasswordChangeForm
+from django.contrib.auth import update_session_auth_hash
+from .forms import UserUpdateForm, ProfileUpdateForm
+from .models import Profile
+
 
 def main(request):
     popular_products = Product.objects.all().order_by('-id')[:6]
@@ -141,6 +148,17 @@ def remove_from_cart(request, item_id):
 def checkout(request):
     cart = get_object_or_404(Cart, user=request.user)
     
+    if request.method == 'POST':
+        address = request.POST.get('address')
+        email = request.POST.get('email') # Забираем введенный email из форм
+        # Валидация: проверяем, что оба поля заполнены
+        if not address or not email:
+            messages.error(request, "Пожалуйста, заполните все обязательные поля!")
+        # Если у самого юзера почты не было, сохраняем её в его профиль
+        if not request.user.email:
+            request.user.email = email
+            request.user.save()
+    
     if not cart.items.exists():
         messages.error(request, "Ваша корзина пуста")
         return redirect('catalog:cart_view')
@@ -172,11 +190,26 @@ def checkout(request):
         )
         email.attach(f'receipt_{cart.id}.xlsx', buffer.getvalue(), 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet')
         email.send()
-
-        cart.items.all().delete()
-        
-        messages.success(request, "Заказ оформлен! Чек отправлен на вашу почту.")
-        return redirect('catalog:product_list')
+        # Цикл для уменьшения количества товаров на складе
+        for item in cart.items.all():
+            product = item.product
+            
+            # 1. Проверяем, существует ли у товара поле остатка (замени 'stock' на свое, если оно называется иначе)
+            if hasattr(product, 'stock') and product.stock is not None:
+                
+                # 2. Вычитаем количество купленного товара из остатка
+                product.stock -= item.quantity
+                
+                # 3. Страховка: если ушли в минус, принудительно ставим 0
+                if product.stock < 0:
+                    product.stock = 0
+                    
+                # 4. Сохраняем обновленный товар обратно в базу данных
+                product.save()
+                cart.items.all().delete()
+                
+                messages.success(request, "Заказ оформлен! Чек отправлен на вашу почту.")
+                return redirect('catalog:product_list')
 
     return render(request, 'shop/checkout.html', {'cart': cart})
 
@@ -217,3 +250,101 @@ class CartItemViewSet(viewsets.ModelViewSet):
     
     def get_queryset(self):
         return CartItem.objects.filter(cart__user=self.request.user)
+    
+def register_view(request):
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save()
+            login(request, user)  # Сразу логиним пользователя после регистрации
+            return redirect('catalog:profile_page')  # Перенаправление в личный кабинет
+    else:
+        form = UserCreationForm()
+    return render(request, 'register.html', {'form': form})
+
+# Вход (Логин)
+def login_view(request):
+    if request.method == 'POST':
+        form = AuthenticationForm(request, data=request.POST)
+        if form.is_valid():
+            username = form.cleaned_data.get('username')
+            password = form.cleaned_data.get('password')
+            user = authenticate(username=username, password=password)
+            if user is not None:
+                login(request, user)
+                return redirect('catalog:profile_page')
+    else:
+        form = AuthenticationForm()
+    return render(request, 'login.html', {'form': form})
+
+# Выход (Логаут)
+def logout_view(request):
+    logout(request)
+    return redirect('catalog:login')
+
+class MyProfileView(generics.RetrieveUpdateAPIView):
+    serializer_class = ProfileSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def get_object(self):
+        return self.request.user.profile
+    
+
+class OrderListView(generics.ListAPIView):
+    serializer_class = OrderSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def get_queryset(self):
+        user = self.request.user
+        if user.profile.role == 'ADMIN':
+            return Order.objects.all()
+        return Order.objects.filter(user=user)
+    
+@login_required
+def profile_page(request):
+    # Автоматически создаем профиль, если его нет в БД для этого юзера
+    profile, created = Profile.objects.get_or_create(user=request.user)
+    
+    # Получаем заказы пользователя (если есть связь)
+    orders = request.user.orders.all() if hasattr(request.user, 'orders') else []
+    
+    return render(request, 'profile.html', {
+        'orders': orders,
+        'profile': profile
+    })
+
+@login_required
+def settings_page(request):
+    profile, created = Profile.objects.get_or_create(user=request.user)
+    
+    if request.method == 'POST':
+        # Если отправлена форма изменения данных
+        if 'update_profile' in request.POST:
+            u_form = UserUpdateForm(request.POST, instance=request.user)
+            p_form = ProfileUpdateForm(request.POST, instance=profile)
+            p_form.fields['favorite_category'].queryset = Category.objects.all()
+            
+            if u_form.is_valid() and p_form.is_valid():
+                u_form.save()
+                p_form.save()
+                messages.success(request, 'Ваш профиль успешно обновлен!')
+                return redirect('catalog:settings_page')
+        
+        # Если отправлена форма смены пароля
+        elif 'change_password' in request.POST:
+            password_form = PasswordChangeForm(request.user, request.POST)
+            if password_form.is_valid():
+                user = password_form.save()
+                update_session_auth_hash(request, user) # Чтобы сессия не слетала
+                messages.success(request, 'Ваш пароль был успешно изменен!')
+                return redirect('catalog:settings_page')
+    else:
+        u_form = UserUpdateForm(instance=request.user)
+        p_form = ProfileUpdateForm(instance=profile)
+        password_form = PasswordChangeForm(request.user)
+
+    return render(request, 'settings.html', {
+        'u_form': u_form,
+        'p_form': p_form,
+        'password_form': password_form
+    })
